rename: 공통컴포넌트 Common 폴더로 이동
